@@ -14,13 +14,13 @@
     "@types/jest": "^27.0.3",
     "@types/node": "^16.11.17",
     "@types/pako": "^2.0.0",
-    "@types/react": "^17.0.38",
-    "@types/react-dom": "^17.0.11",
+    "@types/react": "^18.0.33",
+    "@types/react-dom": "^18.0.11",
     "@vitejs/plugin-react": "^3.1.0",
     "immer": "^9.0.21",
     "pako": "^2.1.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "recoil": "^0.7.7",
     "tar-stream": "npm:tar-stream-browser@^3.0.0",
     "typescript": "^4.5.4",
